Added softmax, huber loss and fixed buffer int storage bug

+++ rlkit/data_management/simple_replay_buffer.py
@@ -13,13 +13,13 @@ class SimpleReplayBuffer(ReplayBuffer):
         self._action_dim = action_dim
         self._max_replay_buffer_size = max_replay_buffer_size
         self._observations = np.zeros(
-            (max_replay_buffer_size, observation_dim), dtype="uint8"
+            (max_replay_buffer_size, observation_dim), dtype="float32"
         )
         # It's a bit memory inefficient to save the observations twice,
         # but it makes the code *much* easier since you no longer have to
         # worry about termination conditions.
         self._next_obs = np.zeros(
-            (max_replay_buffer_size, observation_dim), dtype="uint8"
+            (max_replay_buffer_size, observation_dim), dtype="float32"
         )
         self._actions = np.zeros((max_replay_buffer_size, action_dim), dtype="uint8")
         # Make everything a 2D np array to make it easier for other code to
